% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{balance.sample}
\alias{balance.sample}
\title{Reports a balanced sample of a data frame's factors}
\usage{
balance.sample(df, response, factor.colmn, replace = FALSE,
  min.samp.numb = "min")
}
\arguments{
\item{df}{data frame passed containing value vector and factors}

\item{response}{positive integer index or character name of a column response vector}

\item{factor.colmn}{vector of positive integer(s) index(es) or character name of column(s)
that describes the index of the factors to be used}

\item{replace}{Not yet working}

\item{min.samp.numb}{Can take a positive integer, only subsets of factors
greater than or equal to this number will be reported}
}
\value{
Balanced data frame that is a subset of the first
}
\description{
will return a balancd data frame subset of the original
}
