% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsupliment.R
\name{break.dist}
\alias{break.dist}
\title{returns the distance between numbers}
\usage{
break.dist(breaks)
}
\arguments{
\item{breaks}{used to find distance between numbers}
}
\value{
returns the distance betweens numbers
}
