% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{char.position}
\alias{char.position}
\title{Find positions of a particular pattern.}
\usage{
char.position(x, pattern, position = 0, instance = 1)
}
\arguments{
\item{x}{Character vector to search through}

\item{pattern}{character string that the function will attempt to
find in x}

\item{position}{positive or negative integer used to modulate the position
returned if pattern is found in x, char.position returns the exact position by
default. -1 returns position before pattern and 1 returns position after pattern.}

\item{instance}{positive integer used to modulate which instance of the pattern
is returned. Default is first instance. Assign "last" to return last instance in
string}
}
\value{
Number vector of position of pattern found, returns NA if not found
}
\description{
designed to find the position of a particular character
in a string. Useful with substr function
}
