% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justinmisc.pca.R
\name{circleFun}
\alias{circleFun}
\title{circleFun}
\source{
https://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2/6863490#6863490 from Joran
}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{coordinate points for the center of circle}

\item{diameter}{diameter length of circle, 2r}

\item{npoints}{number of points to estimate}
}
\value{
returns data frame to draw circle points.
}
