% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{clust}
\alias{clust}
\title{standard hclust function}
\usage{
clust(df, col.index, value.var)
}
\arguments{
\item{df}{data frame passed}

\item{col.index}{index or character name of columns in df that will be clustered}

\item{value.var}{value by which clustering occures}
}
\value{
df but with reordered factors
}
\description{
internal clustering function to automatically
reorder factors
}
