% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsupliment.R
\name{color.split}
\alias{color.split}
\title{Splitting color evenly}
\usage{
color.split(start, end, into)
}
\arguments{
\item{start}{color code passed of where to begin}

\item{end}{color code passed of where to end}

\item{into}{number of colors that will be found between start and end}
}
\value{
color vector the size of into with colors evenly spaced apart
}
\description{
Splits the two color hexidecimal codes into evenly spaced colors
}
