% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{comb.comp}
\alias{comb.comp}
\title{generate combinations}
\usage{
comb.comp(vector, rtrn.all = FALSE)
}
\arguments{
\item{vector}{factor vector of at least 2}

\item{rtrn.all}{logical, controlled internally}
}
\value{
dataframe
}
\description{
Generate a dataframe of the 2 group combinations between factors
}
