% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{combo.wilcox}
\alias{combo.wilcox}
\title{Wilcoxon tests through many factors}
\usage{
combo.wilcox(df, response, factor, p.v.correct = FALSE,
  single.factor.comp = FALSE)
}
\arguments{
\item{df}{data frame passed containing value vector and factors}

\item{response}{positive integer index or character name of a column response vector}

\item{p.v.correct}{Logical assignment. Default set to FALSE. Will return logical vector
along with test results. TRUE indicates P.value is less than adjusted P.value, FALSE
indicates P.value is greated than adjusted P.value. Adjusted P.value is calculated using
the bonferroni correction.}

\item{single.factor.comp}{logical assignment. Default set to FALSE. Set to TRUE when you
user wants compare between two levels of the same factor while all other factors are
held constant. When set to TRUE, factor vector must be at least length 2.}

\item{factor.colmn}{vector of positive integer(s) index(es) or character name of column(s)
that describes the index of the factors to be used. Number of factors passed changes the
comparison. One factor vector passed will return wilcox test values between levels. Multiple
factor vectors passed will return wilcon test values between all combination of interactions
of factors.}
}
\value{
dataframe of wilcoxn test P.value results
}
\description{
Calculate the wilcoxon test between several factors
}
