% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{commoncol}
\alias{commoncol}
\title{Find Common Column Names}
\usage{
commoncol(x, y, join = FALSE)
}
\arguments{
\item{x}{data frame}

\item{y}{data frame}

\item{join}{Logical, augments the return. Default set to FALSE. Still
a work in progress}
}
\value{
If join = FALSE this will return character string vector of the
common names. If join = TRUE this will try to return character string in
format used for innerjoin.
}
\description{
Used to find common column names of two data frames.
Useful for large data sets
}
