% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{count.to.frequency}
\alias{count.to.frequency}
\title{reshape count data into frequency data}
\usage{
count.to.frequency(df, count, drop.na = FALSE)
}
\arguments{
\item{df}{data frame}

\item{count}{Column vector name (character) or index (integer)
that specifies count data}

\item{drop.na}{Logical: Default set to FALSE for speed, NA may
affect output thus set to TRUE if unsure.}
}
\value{
df with mode types
}
\description{
Change count data into frequency data by repeating
lines of data frame via the count column.
}
