% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{index.o.coln}
\alias{index.o.coln}
\title{Index or Column name}
\usage{
index.o.coln(vec, v.size, v.name, name.col)
}
\arguments{
\item{vec}{vector we are testing}

\item{v.size}{expected size of vec}

\item{v.name}{argument name of vec in its function call}

\item{name.col}{name of all column vectors}
}
\value{
integer vector of size v.size
}
\description{
Generally internal funciton mostly used to see if vector
 passed specifies column by character name or index
}
