% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justinmisc.pca.R
\name{just.pca}
\alias{just.pca}
\title{just.pca}
\usage{
just.pca(tidydata, y, x, value.var, scale = FALSE, num.degrees = 5,
  invert = FALSE)
}
\arguments{
\item{tidydata}{Tidy data input}

\item{y}{column name(s) or index(es) of columns whose interaction is used
as y component of formula in dcast's formula = y~x. y or the interaction
of columns passed to y becomes the 'individuals' of the pca. This interaction
forms an 'id.col' that is added to the original data frame and used to join
pca components and the original data frame.}

\item{x}{column name(s) or index(es) of columns whose interaction is used
as y component of formula in dcast's formula = y~x. x or the interaction
of columns passed to x becomes the 'variables' of the pca.}

\item{value.var}{Character string of name of column which stores values.}

\item{scale}{Logical, scales the variables to have the same unit variance.
Default set to FALSE.}

\item{num.degrees}{Integer that determines the number of principal components
to join to the orginal data frame. Default set to 5.}

\item{invert}{Logical, determines if mean centering should be inverted or not.
Default set to FALSE.}
}
\value{
returns pca object that is a list of "data","loadings","values",
"covariance","contribution"
}
\description{
Takes Tidy data and Generates pca.object that retains
information of original data.
}
