% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{lm.test}
\alias{lm.test}
\title{linear model testing}
\usage{
lm.test(df, factors, formula)
}
\arguments{
\item{df}{data frame}

\item{factors}{factor column names or indexs to subset by}

\item{formula}{formula to use in lm function}
}
\value{
vector of r-squared values
}
\description{
report r-squared values of subsets
}
