% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{make.atomic}
\alias{make.atomic}
\title{Make data frame atomic}
\usage{
make.atomic(df, de.atomic, sep, new.name = "atomic")
}
\arguments{
\item{df}{data frame}

\item{de.atomic}{Index of the column or character column
name to make atomic}

\item{sep}{character pattern delimiter for nested factors}

\item{new.name}{Name of the atomic column}
}
\value{
denested data frame
}
\description{
Used when one column has nested values that
make the data frame not atomic
}
