% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{merge_multi}
\alias{merge_multi}
\title{merge_multi}
\usage{
merge_multi(...)
}
\arguments{
\item{...}{vectors to be merged. argument names become column names in returned value.
If no arguments specified, default column names set to LETTERS set.}
}
\value{
data frame of combinded vectors.
}
\description{
extends merge function to 3 or more vectors
}
\examples{

merge_multi(x = c(1,2,3), y = c(4,5,6), z = c(7,8,9))
merge_multi(c(1,2,3),c(4,5,6),c(7,8,9))

}
