% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{mode.type}
\alias{mode.type}
\title{Mode Type}
\usage{
mode.type(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Return character vector of mode types
}
\description{
Find the mode of the variables of a df as
dictated by str
}
