% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{paste2}
\alias{paste2}
\title{Paste that removes NAs}
\source{
https://stackoverflow.com/questions/13673894/suppress-nas-in-paste
}
\usage{
paste2(..., sep = ", ")
}
\arguments{
\item{sep}{character to separate paste2 arguments}
}
\value{
Character string
}
\description{
Use this Function to ignore NAs in paste
}
