% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justinmisc.pca.R
\name{pca.biplot}
\alias{pca.biplot}
\title{pca.biplot}
\usage{
pca.biplot(pca.obj, pcs, flip = FALSE)
}
\arguments{
\item{pca.obj}{Takes pca object made by just.pca()}

\item{pcs}{Integer input vector of length 2 of which PC dimensions
should be plotted}

\item{flip}{Logical, Defaul set to FALSE. will flip coordinates
over x=y line. Only included because PCA function's biplot seems to
always be flipped.}
}
\value{
ggplot plot of the variables in a coordinate graph of two
pc dimensions.
}
\description{
Returns biplot of two pc dimensions
}
