% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justinmisc.pca.R
\name{plotContrib}
\alias{plotContrib}
\title{plotContrib}
\usage{
plotContrib(pca.obj, PC, first = 10)
}
\arguments{
\item{pca.obj}{Takes pca object made by just.pca()}

\item{PC}{Integer input of which PC dimension should be plotted}

\item{first}{Integer input. Returns the first 10 most contributing
variables by default.}
}
\value{
ggplot bar plot of the the most contributing variables to
a PC dimension.
}
\description{
plots the contribution of variables in pca in ggplot
}
