% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{reorderFactor.by}
\alias{reorderFactor.by}
\title{Reorder factors according to simple funciton}
\usage{
reorderFactor.by(data, factor.to.order, factor.col, by.factor, FUN = NULL,
  applied.on, group.by = NULL, decreasing = FALSE)
}
\arguments{
\item{data}{data frame}

\item{factor.to.order}{column name or index of factors that will be reordered}

\item{factor.col}{column name or index that contains 'by.factor'}

\item{by.factor}{character name of factor in factor.col that data will
be subsetted for calculations and group desired for order}

\item{FUN}{Simple function that takes a vector of numbers and returns a single
output. Default set to NULL, must assign a function if group.by is not NULL.}

\item{applied.on}{Numeric vector used as inputs to FUN. Will ultamitly
determine order of factor.to.order}

\item{group.by}{column names or indexes that further groups by.factor subset to
which FUN is applied. Default set to NULL.}

\item{decreasing}{logical, default set to FALSE. Determines if order is increasing
or decreasing.}
}
\value{
returns data with factor.to.order reordered by FUN of by.factor subsetted through group.by.
}
\description{
Will reorder factors based off of a group subset
}
