% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justinmisc.pca.R
\name{to.unit}
\alias{to.unit}
\title{to.unit}
\usage{
to.unit(x)
}
\arguments{
\item{x}{numeric vector passed of length 2}
}
\value{
returns x y cordinates in polar form
}
\description{
Takes cartesian coordiants and returns the unit circle
points as x, y coordinates.
}
