% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{twokeffect}
\alias{twokeffect}
\title{perform 2^k analysis on factors}
\usage{
twokeffect(df, response, factor.colmn, interaction = F)
}
\arguments{
\item{df}{data frame passed containing value vector and factors}

\item{response}{positive integer index or character name of a column response vector}

\item{factor.colmn}{vector of positive integer(s) index(es) or character name of column(s)
that describes the index of the factors to be used}

\item{interaction}{Logical, augments if interaction between factor levels will
be reported. Default is set to FALSE}
}
\value{
data frame of effects
}
\description{
will report the effects of various factors of two levels
}
