% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{correct.mode}
\alias{correct.mode}
\title{correct mode types}
\usage{
correct.mode(df, mode.vec, space_replace = "_")
}
\arguments{
\item{df}{data frame}

\item{mode.vec}{designed to take output of function mode.type(),
mode.vec must be one of the following: "num","int","Factor",
"cplx","chr","logi","raw", "ignore". Mode type is changed by index.}

\item{space_replace}{character to replace " " characters in column names.}
}
\value{
df with mode types
}
\description{
Function to correct modes after df
manipulation changes columns to characters
}
