% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfanalysis.R
\name{distance.cluster}
\alias{distance.cluster}
\title{euclidian clustering function}
\usage{
distance.cluster(df, formula, value.var, which.clust = "both",
  fun.aggregate = mean)
}
\arguments{
\item{df}{data frame passed containing both factor columns to reorder
and numeric column to cluster by}

\item{value.var}{Character column name by which clustering will be
calculated}

\item{y.by.x}{Index of the columns or character column
names to reorder}

\item{which.clut}{Defines which column is reordered. Default set to both.
Set to "y" to only reorder y.by.x[1], set to "x" to only reorder y.by.x[2]}
}
\value{
returns df but with reordered factors
}
\description{
reorder factors passed to y.by.x via their respective
clusters
}
